diplomat_runtime::write

Function diplomat_simple_write

#[unsafe(no_mangle)]
pub unsafe extern "C" fn diplomat_simple_write(
    buf: *mut u8,
    buf_size: usize,
) -> DiplomatWrite

Create an DiplomatWrite that can write to a fixed-length stack allocated u8 buffer.

Once done, this will append a null terminator to the written string.

This is largely used internally by Diplomat-generated FFI code, and should not need to be constructed manually outside of that context. See RustWriteVec if you need this in Rust.

Safety

buf must be a valid pointer to a region of memory that can hold at buf_size bytes