diplomat_runtime

Module writeable

Structs

DiplomatWriteable

An object that can one can write UTF-8 strings to

Functions

diplomat_buffer_writeable_create

Create an DiplomatWriteable that can write to a dynamically allocated buffer managed by Rust.

diplomat_buffer_writeable_destroy⚠

Destructor for Rust-memory backed writables.

diplomat_buffer_writeable_get_bytes

Grabs a pointer to the underlying buffer of a writable.

diplomat_buffer_writeable_len

Gets the length in bytes of the content written to the writable.

diplomat_simple_writeable⚠

Create an DiplomatWriteable that can write to a fixed-length stack allocated u8 buffer.