Crate diplomat_runtime

Modules

result 🔒

writeable 🔒

Structs

DiplomatResult

A Result-like type that can be passed across the FFI boundary as a value. Used internally to return Results and Options from functions.

DiplomatWriteable

An object that can one can write UTF-8 strings to

Functions

diplomat_alloc⚠

Allocates a buffer of a given size in Rust’s memory.

diplomat_free⚠

Frees a buffer that was allocated in Rust’s memory.

Type Aliases

DiplomatByte

Like u8, but interpreted explicitly as a raw byte as opposed to a numerical value. This matters for languages like JavaScript or Dart, where there’s only a single numeric type, but special types for byte buffers.

DiplomatChar

Like char, but unvalidated.

DiplomatStr

Like str, but unvalidated.

DiplomatStr16

Like Wstr, but unvalidated.